Add admin urls
