+++ frontend/src/containers/pages/accountability/communication/CreateCommunicationPage.tsx
@@ -46,6 +46,8 @@ import { FormikSliderField } from '@shared/Formik/FormikSliderField';
 import { FormikTextField } from '@shared/Formik/FormikTextField';
 import { CommunicationSteps, CommunicationTabsValues } from '@utils/constants';
 import { getPercentage } from '@utils/utils';
+import { GenericAdminButton } from '@core/AdminButton';
+import * as React from 'react';
 
 const steps = ['Recipients Look up', 'Sample Size', 'Details'];
 const SampleSizeTabs = ['Full List', 'Random Sampling'];
@@ -330,6 +332,7 @@ export const CreateCommunicationPage = (): React.ReactElement => {
                 ? breadCrumbsItems
                 : null
             }
+            flags={<GenericAdminButton currentUrl={document.location.href} />}
           />
           <PaperContainer>
             <Grid xs={9} item>
