+++ frontend/src/utils/utils.ts
@@ -1052,56 +1052,3 @@ export const isProgramNodeUuidFormat = (id: string): boolean => {
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export const arraysHaveSameContent = (a: any[], b: any[]): boolean =>
   a.length === b.length && a.every((val, index) => val === b[index]);
-
-export const generateAdminRedirect = (url: string) => {
-    const urlObject = new URL(url);
-    const urlPathnameElements = urlObject.pathname.split('/');
-    const encodedUUID = urlPathnameElements.pop();
-
-    if (urlPathnameElements.length === 1) {
-      return null;
-    }
-
-    let decodedUUID;
-    try {
-      decodedUUID = atob(encodedUUID).split(':')[1];
-    } catch (e) {
-      return null;
-    }
-
-    const urlPathSwitch = urlPathnameElements.slice(4).join('/');
-    const adminBaseUrl = `${urlObject.origin}/api/unicorn`;
-
-    switch (urlPathSwitch) {
-      case 'registration-data-import':
-        return `${adminBaseUrl}/registration_data/registrationdataimport/${decodedUUID}/change`;
-      case 'population/household':
-        return `${adminBaseUrl}/household/household/${decodedUUID}/change`;
-      case 'population/individuals':
-        return `${adminBaseUrl}/household/individual/${decodedUUID}/change`;
-      case 'details':
-        return `${adminBaseUrl}/program/program/${decodedUUID}/change`;
-      case 'target-population':
-        return `${adminBaseUrl}/targeting/targetpopulation/${decodedUUID}/change`;
-      case 'payment-module/payment-plans':
-        return `${adminBaseUrl}/payment/paymentplan/${decodedUUID}/change`;
-      case 'payment-module/payments':
-        return `${adminBaseUrl}/payment/payment/${decodedUUID}/change`;
-      case 'payment-verification/payment-plan':
-        return `${adminBaseUrl}/payment/paymentverificationplan/${decodedUUID}/change`;
-      case 'verification/payment':
-        return `${adminBaseUrl}/payment/paymentverification/${decodedUUID}/change`;
-      case 'grievance/tickets/user-generated':
-        return `${adminBaseUrl}/grievance/grievanceticket/${decodedUUID}/change`;
-      case 'grievance/tickets/system-generated':
-        return `${adminBaseUrl}/grievance/grievanceticket/${decodedUUID}/change`;
-      case 'grievance/feedback':
-        return `${adminBaseUrl}/accountability/feedback/${decodedUUID}/change`;
-      case 'accountability/communication':
-        return `${adminBaseUrl}/accountability/message/${decodedUUID}/change`;
-      case 'accountability/surveys':
-        return `${adminBaseUrl}/accountability/survey/${decodedUUID}/change`;
-      default:
-        return null;
-    }
-  };
